// src/app/models/securityTransaction.ts
export class SecurityTransaction {
    date: Date;
    strDate:string;
    high: number;
    low: number;
    openPrice: number;
    closePrice: number;
    tickerName: string;
    ticker: string;
    scenario : string;
    indexName: string;
    rowNum: number;
    topBottom: number;
    isHigh: number;
}

export class otReport{
    
}